requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
openai>=1.40.0
httpx>=0.27.0
PyYAML>=6.0.0
requests>=2.31.0
pytest>=8.0.0
json-repair>=0.59.0
